+++ ICD.Connect.Telemetry.Crestron/Assets/IFusionAssetDataCollection.cs
@@ -10,13 +10,5 @@ namespace ICD.Connect.Telemetry.Crestron.Assets
 		/// <param name="id"></param>
 		/// <returns></returns>
 		FusionAssetData this[uint id] { get; }
-
-		/// <summary>
-		/// Gets the first asset data with the given asset type.
-		/// </summary>
-		/// <param name="type"></param>
-		/// <returns></returns>
-		[CanBeNull]
-		FusionAssetData GetAssetData(eAssetType type);
 	}
 }