+++ ICD.Connect.Telemetry.Crestron/Devices/MockFusionRoom/MockFusionRoomSettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Devices;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Telemetry.Crestron.Devices
+{
+	[KrangSettings("MockFusionRoom", typeof(MockFusionRoom.MockFusionRoom))]
+	public sealed class MockFusionRoomSettings : AbstractDeviceSettings
+	{
+	}
+}